Fix default job store selection and error messaging for grid batch systems
