+++ src/toil/worker.py
@@ -34,7 +34,7 @@ from typing import Any
 
 from configargparse import ArgParser
 
-from toil import logProcessContext
+from toil import logProcessContext, options
 from toil.common import Config, Toil, safeUnpickleFromStream
 from toil.cwl.utils import (
     CWL_UNSUPPORTED_REQUIREMENT_EXCEPTION,
@@ -48,7 +48,7 @@ from toil.job import (
     Job,
     JobDescription,
 )
-from toil.jobStores.abstractJobStore import AbstractJobStore
+from toil.jobStores.abstractJobStore import AbstractJobStore, NoSuchJobStoreException
 from toil.lib.io import make_public_dir, path_union
 from toil.lib.resources import ResourceMonitor
 from toil.statsAndLogging import StatsDict, configure_root_logger, install_log_color, set_log_level
@@ -985,7 +985,17 @@ def main(argv: list[str] | None = None) -> None:
     # Load the jobStore/config file
     ##########################################
 
-    job_store = Toil.resumeJobStore(options.jobStoreLocator)
+    try:
+        job_store = Toil.resumeJobStore(options.jobStoreLocator)
+    except NoSuchJobStoreException:
+        raise RuntimeError(
+            f"Could not access the job store at '{options.jobStoreLocator}'. "
+            f"If you are running with a grid batch system (e.g. Slurm), the job store "
+            f"path must be on a shared filesystem accessible from all worker nodes. "
+            f"Use --jobStore to point to shared storage (e.g. NFS, Lustre) instead of "
+            f"a local path."
+        )
+    
     config = job_store.config
 
     with in_contexts(options.context):
